feat: refactor font handling
