+++ apps/web/src/features/tui/terminal-preferences.ts
@@ -3,22 +3,6 @@ import { persist } from 'zustand/middleware'
 
 import { persistStorage } from '~/store/persist-storage'
 
-export const DEFAULT_TERMINAL_FONT_FAMILY = [
-  '"Geist Mono"',
-  '"GeistMono"',
-  '"Cascadia Code"',
-  '"Cascadia Mono"',
-  '"Fira Mono"',
-  '"MesloLGM Nerd Font Mono"',
-  '"CaskaydiaCove Nerd Font"',
-  '"CaskaydiaMono Nerd Font"',
-  '"FiraCode Nerd Font Mono"',
-  '"Symbols Nerd Font Mono"',
-  '"PowerlineSymbols"',
-  'ui-monospace',
-  'monospace',
-].join(', ')
-
 export interface TerminalPreferencesState {
   fontFamily: string | null
   setFontFamily: (fontFamily: string | null | undefined) => void
@@ -28,11 +12,6 @@ export interface TerminalPreferencesState {
 export function normalizeTerminalFontFamily(fontFamily: string | null | undefined): string | null {
   const trimmed = fontFamily?.trim() ?? ''
   return trimmed.length > 0 ? trimmed : null
-}
-
-export function resolveTerminalFontFamily(fontFamily: string | null | undefined): string {
-  const normalized = normalizeTerminalFontFamily(fontFamily)
-  return normalized ? `${normalized}, ${DEFAULT_TERMINAL_FONT_FAMILY}` : DEFAULT_TERMINAL_FONT_FAMILY
 }
 
 export const useTerminalPreferencesStore = create<TerminalPreferencesState>()(
