+++ apps/web/src/locales/zh-CN/settings.json
@@ -50,7 +50,7 @@
   "appearance.sessionPreview.label": "折叠时会话预览数",
   "appearance.terminalFont.description": "此设备内置终端使用的字体族。留空则使用 Cradle 默认值。",
   "appearance.terminalFont.label": "终端字体",
-  "appearance.terminalFont.placeholder": "MesloLGM Nerd Font Mono, CaskaydiaCove Nerd Font",
+  "appearance.terminalFont.placeholder": "\"MesloLGM Nerd Font Mono\", \"CaskaydiaCove Nerd Font\", ui-monospace, monospace",
   "appearance.theme.dark": "深色",
   "appearance.theme.description": "选择应用的颜色主题。",
   "appearance.theme.info": "浅色和深色主题适合不同环境。系统模式会跟随操作系统偏好。",
